Default to PWD for workdir now that we are a bin
Now that we're no longer only executing from with bazel we can take
`--cwd` as an optional pramater. This lets us operate like
`docker-compose` does where you just type `docker-compose up` and it
operates as if your CWD is the root.

TODO(josh): search through directory tree for WORKSPACE starting from
CWD and moving upward. This way it's *really* as usable as
`docker-compose`.